requirements.txt
nextcord>=3.1.1
httpx>=0.28.1
python-dotenv>=1.1.1
elevenlabs>=1.0.0
